requirements.txt
aiofiles==23.2.1
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
colorama==0.4.6
fake-useragent==1.2.1
greenlet==2.0.2
h11==0.14.0
html-parser==0.2
html5==0.0.9
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
loguru==0.7.0
lxml==4.9.3
mysql==0.0.3
mysqlclient==2.2.0
numpy==1.25.2
pandas==2.0.3
playwright==1.37.0
playwright-stealth==1.0.5
ply==3.11
pyee==9.0.4
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
webencodings==0.5.1
win32-setctime==1.1.0
